+++ tests/deprecated_api/test_remove_1-6.py
@@ -17,6 +17,7 @@ import pytest
 from pytorch_lightning import Trainer
 from pytorch_lightning.callbacks.early_stopping import EarlyStopping
 from pytorch_lightning.plugins.training_type import DDPPlugin, DDPSpawnPlugin
+from pytorch_lightning.utilities.model_helpers import is_overridden
 from tests.helpers import BoringDataModule, BoringModel
 
 
@@ -173,6 +174,14 @@ def test_v1_6_0_datamodule_hooks_calls(tmpdir):
     assert dm.prepare_data_calls == 1
     assert dm.setup_calls == ['fit', None]
     assert dm.teardown_calls == ['validate', 'test']
+
+
+def test_v1_6_0_is_overridden_model():
+    model = BoringModel()
+    with pytest.deprecated_call(match="and will be removed in v1.6"):
+        assert is_overridden("validation_step", model=model)
+    with pytest.deprecated_call(match="and will be removed in v1.6"):
+        assert not is_overridden("foo", model=model)
 
 
 def test_v1_6_0_early_stopping_monitor(tmpdir):